+++ protocols/kad/src/handler.rs
@@ -22,27 +22,24 @@ use crate::behaviour::Mode;
 use crate::protocol::{
     KadInStreamSink, KadOutStreamSink, KadPeer, KadRequestMsg, KadResponseMsg, ProtocolConfig,
 };
-use crate::record_priv::{self, Record};
+use crate::record::{self, Record};
 use crate::QueryId;
 use either::Either;
 use futures::prelude::*;
 use futures::stream::SelectAll;
-use instant::Instant;
 use libp2p_core::{upgrade, ConnectedPoint};
 use libp2p_identity::PeerId;
 use libp2p_swarm::handler::{
     ConnectionEvent, DialUpgradeError, FullyNegotiatedInbound, FullyNegotiatedOutbound,
 };
 use libp2p_swarm::{
-    ConnectionHandler, ConnectionHandlerEvent, ConnectionId, KeepAlive, Stream, StreamUpgradeError,
+    ConnectionHandler, ConnectionHandlerEvent, ConnectionId, Stream, StreamUpgradeError,
     SubstreamProtocol, SupportedProtocols,
 };
 use log::trace;
 use std::collections::VecDeque;
 use std::task::Waker;
-use std::{
-    error, fmt, io, marker::PhantomData, pin::Pin, task::Context, task::Poll, time::Duration,
-};
+use std::{error, fmt, io, marker::PhantomData, pin::Pin, task::Context, task::Poll};
 
 const MAX_NUM_SUBSTREAMS: usize = 32;
 
@@ -60,9 +57,6 @@ pub struct Handler {
     /// In client mode, we don't accept inbound substreams.
     mode: Mode,
 
-    /// Time after which we close an idle connection.
-    idle_timeout: Duration,
-
     /// Next unique ID of a connection.
     next_connec_unique_id: UniqueConnecId,
 
@@ -78,9 +72,6 @@ pub struct Handler {
 
     /// List of active inbound substreams with the state they are in.
     inbound_substreams: SelectAll<InboundSubstreamState>,
-
-    /// Until when to keep the connection alive.
-    keep_alive: KeepAlive,
 
     /// The connected endpoint of the connection that the handler
     /// is associated with.
@@ -235,7 +226,7 @@ pub enum HandlerEvent {
     /// this key.
     GetProvidersReq {
         /// The key for which providers are requested.
-        key: record_priv::Key,
+        key: record::Key,
         /// Identifier of the request. Needs to be passed back when answering.
         request_id: RequestId,
     },
@@ -261,7 +252,7 @@ pub enum HandlerEvent {
     /// The peer announced itself as a provider of a key.
     AddProvider {
         /// The key for which the peer is a provider of the associated value.
-        key: record_priv::Key,
+        key: record::Key,
         /// The peer that is the provider of the value for `key`.
         provider: KadPeer,
     },
@@ -269,7 +260,7 @@ pub enum HandlerEvent {
     /// Request to get a value from the dht records
     GetRecord {
         /// Key for which we should look in the dht
-        key: record_priv::Key,
+        key: record::Key,
         /// Identifier of the request. Needs to be passed back when answering.
         request_id: RequestId,
     },
@@ -294,7 +285,7 @@ pub enum HandlerEvent {
     /// Response to a request to store a record.
     PutRecordRes {
         /// The key of the stored record.
-        key: record_priv::Key,
+        key: record::Key,
         /// The value of the stored record.
         value: Vec<u8>,
         /// The user data passed to the `PutValue`.
@@ -386,7 +377,7 @@ pub enum HandlerIn {
     /// this key.
     GetProvidersReq {
         /// Identifier being searched.
-        key: record_priv::Key,
+        key: record::Key,
         /// Custom user data. Passed back in the out event when the results arrive.
         query_id: QueryId,
     },
@@ -409,7 +400,7 @@ pub enum HandlerIn {
     /// succeeded.
     AddProvider {
         /// Key for which we should add providers.
-        key: record_priv::Key,
+        key: record::Key,
         /// Known provider for this key.
         provider: KadPeer,
     },
@@ -417,7 +408,7 @@ pub enum HandlerIn {
     /// Request to retrieve a record from the DHT.
     GetRecord {
         /// The key of the record.
-        key: record_priv::Key,
+        key: record::Key,
         /// Custom data. Passed back in the out event when the results arrive.
         query_id: QueryId,
     },
@@ -442,7 +433,7 @@ pub enum HandlerIn {
     /// Response to a `PutRecord`.
     PutRecordRes {
         /// Key of the value that was put.
-        key: record_priv::Key,
+        key: record::Key,
         /// Value that was put.
         value: Vec<u8>,
         /// Identifier of the request that was made by the remote.
@@ -465,7 +456,6 @@ struct UniqueConnecId(u64);
 impl Handler {
     pub fn new(
         protocol_config: ProtocolConfig,
-        idle_timeout: Duration,
         endpoint: ConnectedPoint,
         remote_peer_id: PeerId,
         mode: Mode,
@@ -484,13 +474,9 @@ impl Handler {
             }
         }
 
-        #[allow(deprecated)]
-        let keep_alive = KeepAlive::Until(Instant::now() + idle_timeout);
-
         Handler {
             protocol_config,
             mode,
-            idle_timeout,
             endpoint,
             remote_peer_id,
             next_connec_unique_id: UniqueConnecId(0),
@@ -498,7 +484,6 @@ impl Handler {
             outbound_substreams: Default::default(),
             num_requested_outbound_streams: 0,
             pending_messages: Default::default(),
-            keep_alive,
             protocol_status: None,
             remote_supported_protocols: Default::default(),
             connection_id,
@@ -717,10 +702,6 @@ impl ConnectionHandler for Handler {
         }
     }
 
-    fn connection_keep_alive(&self) -> KeepAlive {
-        self.keep_alive
-    }
-
     fn poll(
         &mut self,
         cx: &mut Context<'_>,
@@ -768,20 +749,6 @@ impl ConnectionHandler for Handler {
                 protocol: SubstreamProtocol::new(self.protocol_config.clone(), ()),
             });
         }
-
-        let no_streams = self.outbound_substreams.is_empty() && self.inbound_substreams.is_empty();
-
-        self.keep_alive = {
-            #[allow(deprecated)]
-            match (no_streams, self.keep_alive) {
-                // No open streams. Preserve the existing idle timeout.
-                (true, k @ KeepAlive::Until(_)) => k,
-                // No open streams. Set idle timeout.
-                (true, _) => KeepAlive::Until(Instant::now() + self.idle_timeout),
-                // Keep alive for open streams.
-                (false, _) => KeepAlive::Yes,
-            }
-        };
 
         Poll::Pending
     }