+++ swarm/CHANGELOG.md
@@ -1,11 +1,21 @@
 ## 0.44.0 - unreleased
 
+- Remove `handler` field from `ConnectionClosed`.
+  If you need to transfer state from a `ConnectionHandler` to its `NetworkBehaviour` when a connection closes, use `ConnectionHandler::poll_close`.
+  See [PR 4076](https://github.com/libp2p/rust-libp2p/pull/4076).
+- Remove deprecated `PollParameters` from `NetworkBehaviour::poll` function.
+  See [PR 4490](https://github.com/libp2p/rust-libp2p/pull/4490).
 - Add `PeerCondition::DisconnectedAndNotDialing` variant, combining pre-existing conditions.
   This is the new default.
   A new dialing attempt is iniated _only if_ the peer is both considered disconnected and there is currently no ongoing dialing attempt.
   See [PR 4225](https://github.com/libp2p/rust-libp2p/pull/4225).
 - Remove deprecated `keep_alive_timeout` in `OneShotHandlerConfig`.
   See [PR 4677](https://github.com/libp2p/rust-libp2p/pull/4677).
+- Don't close entire connection upon `DialUpgradeError`s within `OneShotHandler`.
+  Instead, the error is reported as `Err(e)` via `ConnectionHandler::ToBehaviour`.
+  See [PR 4715](https://github.com/libp2p/rust-libp2p/pull/4715).
+- Log `PeerId` of `Swarm` even when constructed with new `SwarmBuilder`.
+  See [PR 4671](https://github.com/libp2p/rust-libp2p/pull/4671).
 
 ## 0.43.6